html/semantics/scripting-1/the-script-element/scriptexecute-events/005.html

<!DOCTYPE html>
<html>
  <title>beforescriptexecute for document.write generated script</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
  <script>
    targets = [];
    var t = async_test();
    document.addEventListener("beforescriptexecute", function(e) {targets.push(e.target);}, false);
    document.write("<script>t.step(function() {assert_array_equals(targets, [document.getElementsByTagName('script')[3]]);}); t.done();<\/script>");
 </script>
 <script>
   //this just causes us to fail if the document.write script doesn't run at all
   t.step(function(){assert_unreached()})
 </script>
